refactor(gui): extract ego canvas state into useEgoCanvas, stop fitView from fighting it
GraphView had grown past the 600-line complexity budget, so the canvas state
machine (focus, accumulated shown/expanded sets, focus history, recenter) moves
into its own hook.

Doing that surfaced two real defects the fixture had been hiding:

- The fitView prop re-fit the viewport after nodes were measured, overwriting
  the setCenter that puts the focus node in the middle of the pane. The demo
  graph happened to be symmetric about the origin so a bbox fit landed in the
  same place; an asymmetric graph would have slid the focus card back under the
  filter panel and reset the zoom.
- ReactFlow's default double-click zoom consumed the gesture before
  onNodeDoubleClick ran, so double-click-to-recenter never worked at all.
